Migrate files to Dalamud.Logging namespace Remove Log.LogX methods.
