+++ Dalamud/Interface/Internal/InterfaceManager.cs
@@ -11,6 +11,7 @@ using Dalamud.Game.ClientState;
 using Dalamud.Game.Internal.DXGI;
 using Dalamud.Hooking;
 using Dalamud.Hooking.Internal;
+using Dalamud.Logging.Internal;
 using ImGuiNET;
 using ImGuiScene;
 using Serilog;
