+++ Dalamud/Interface/Internal/Windows/PluginInstallerWindow.cs
@@ -13,6 +13,7 @@ using CheapLoc;
 using Dalamud.Interface.Colors;
 using Dalamud.Interface.Components;
 using Dalamud.Interface.Windowing;
+using Dalamud.Logging.Internal;
 using Dalamud.Plugin;
 using Dalamud.Plugin.Internal;
 using Dalamud.Plugin.Internal.Exceptions;
