+++ Dalamud/Logging/PluginLog.cs
@@ -1,0 +1,113 @@
+using System;
+using System.Reflection;
+
+namespace Dalamud.Logging
+{
+    /// <summary>
+    /// Class offering various static methods to allow for logging in plugins.
+    /// </summary>
+    public static class PluginLog
+    {
+        /// <summary>
+        /// Log a templated verbose message to the in-game debug log.
+        /// </summary>
+        /// <param name="messageTemplate">The message template.</param>
+        /// <param name="values">Values to log.</param>
+        public static void Verbose(string messageTemplate, params object[] values)
+            => Serilog.Log.Verbose($"[{Assembly.GetCallingAssembly().GetName().Name}] {messageTemplate}", values);
+
+        /// <summary>
+        /// Log a templated verbose message to the in-game debug log.
+        /// </summary>
+        /// <param name="exception">The exception that caused the error.</param>
+        /// <param name="messageTemplate">The message template.</param>
+        /// <param name="values">Values to log.</param>
+        public static void Verbose(Exception exception, string messageTemplate, params object[] values)
+            => Serilog.Log.Verbose(exception, $"[{Assembly.GetCallingAssembly().GetName().Name}] {messageTemplate}", values);
+
+        /// <summary>
+        /// Log a templated debug message to the in-game debug log.
+        /// </summary>
+        /// <param name="messageTemplate">The message template.</param>
+        /// <param name="values">Values to log.</param>
+        public static void Debug(string messageTemplate, params object[] values)
+            => Serilog.Log.Debug($"[{Assembly.GetCallingAssembly().GetName().Name}] {messageTemplate}", values);
+
+        /// <summary>
+        /// Log a templated debug message to the in-game debug log.
+        /// </summary>
+        /// <param name="exception">The exception that caused the error.</param>
+        /// <param name="messageTemplate">The message template.</param>
+        /// <param name="values">Values to log.</param>
+        public static void Debug(Exception exception, string messageTemplate, params object[] values)
+            => Serilog.Log.Debug(exception, $"[{Assembly.GetCallingAssembly().GetName().Name}] {messageTemplate}", values);
+
+        /// <summary>
+        /// Log a templated information message to the in-game debug log.
+        /// </summary>
+        /// <param name="messageTemplate">The message template.</param>
+        /// <param name="values">Values to log.</param>
+        public static void Information(string messageTemplate, params object[] values)
+            => Serilog.Log.Information($"[{Assembly.GetCallingAssembly().GetName().Name}] {messageTemplate}", values);
+
+        /// <summary>
+        /// Log a templated information message to the in-game debug log.
+        /// </summary>
+        /// <param name="exception">The exception that caused the error.</param>
+        /// <param name="messageTemplate">The message template.</param>
+        /// <param name="values">Values to log.</param>
+        public static void Information(Exception exception, string messageTemplate, params object[] values)
+            => Serilog.Log.Information(exception, $"[{Assembly.GetCallingAssembly().GetName().Name}] {messageTemplate}", values);
+
+        /// <summary>
+        /// Log a templated warning message to the in-game debug log.
+        /// </summary>
+        /// <param name="messageTemplate">The message template.</param>
+        /// <param name="values">Values to log.</param>
+        public static void Warning(string messageTemplate, params object[] values)
+            => Serilog.Log.Warning($"[{Assembly.GetCallingAssembly().GetName().Name}] {messageTemplate}", values);
+
+        /// <summary>
+        /// Log a templated warning message to the in-game debug log.
+        /// </summary>
+        /// <param name="exception">The exception that caused the error.</param>
+        /// <param name="messageTemplate">The message template.</param>
+        /// <param name="values">Values to log.</param>
+        public static void Warning(Exception exception, string messageTemplate, params object[] values)
+            => Serilog.Log.Warning(exception, $"[{Assembly.GetCallingAssembly().GetName().Name}] {messageTemplate}", values);
+
+        /// <summary>
+        /// Log a templated error message to the in-game debug log.
+        /// </summary>
+        /// <param name="messageTemplate">The message template.</param>
+        /// <param name="values">Values to log.</param>
+        public static void Error(string messageTemplate, params object[] values)
+            => Serilog.Log.Error($"[{Assembly.GetCallingAssembly().GetName().Name}] {messageTemplate}", values);
+
+        /// <summary>
+        /// Log a templated error message to the in-game debug log.
+        /// </summary>
+        /// <param name="exception">The exception that caused the error.</param>
+        /// <param name="messageTemplate">The message template.</param>
+        /// <param name="values">Values to log.</param>
+        public static void Error(Exception exception, string messageTemplate, params object[] values)
+            => Serilog.Log.Error(exception, $"[{Assembly.GetCallingAssembly().GetName().Name}] {messageTemplate}", values);
+
+        /// <summary>
+        /// Log a templated fatal message to the in-game debug log.
+        /// </summary>
+        /// <param name="messageTemplate">The message template.</param>
+        /// <param name="values">Values to log.</param>
+        public static void Fatal(string messageTemplate, params object[] values)
+            => Serilog.Log.Fatal($"[{Assembly.GetCallingAssembly().GetName().Name}] {messageTemplate}", values);
+
+        /// <summary>
+        /// Log a templated fatal message to the in-game debug log.
+        /// </summary>
+        /// <param name="exception">The exception that caused the error.</param>
+        /// <param name="messageTemplate">The message template.</param>
+        /// <param name="values">Values to log.</param>
+        public static void Fatal(Exception exception, string messageTemplate, params object[] values)
+            => Serilog.Log.Fatal(exception, $"[{Assembly.GetCallingAssembly().GetName().Name}] {messageTemplate}", values);
+    }
+}
