+++ Dalamud/Plugin/Internal/PluginRepository.cs
@@ -1,9 +1,9 @@
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
-using System.Linq;
 using System.Net;
 using System.Threading.Tasks;
 
+using Dalamud.Logging.Internal;
 using Dalamud.Plugin.Internal.Types;
 using Newtonsoft.Json;
